NMS-1341: Modify example codes and write readme
