+++ requirements.txt
@@ -1,4 +1,5 @@
 pyyaml==5.4.1
 tensorflow>=2.3.1
 numpy==1.19.4
-opencv-python==4.1.1.26
+opencv-python==4.6.0.66
+openvino
